refactor(core): extract runner-file HTTP ops from daemon god file
daemon.rs was a 3388-line god file (audit god_file). Extract the
runner-file HTTP operation cluster — create workspace file directory,
upload/download runner files, and the workspace-path resolve/normalize/
canonicalize helpers — into a runner_files.rs submodule.

Convert daemon.rs to a directory module (git mv preserves history; module
path 'daemon' unchanged, so mod.rs's own crate:: paths stay; fixed the
#[path] to the external daemon_test.rs, now one level deeper). The three
entry points (create/upload/download_runner_file) are pub(super) for the
route handlers; the request structs (FilePathRequest, FileUploadRequest)
are pub(super). External contract unchanged.

Behavior-preserving: daemon/mod.rs 3388 -> 3197 lines, runner_files.rs
208. Runner-file route tests pass. (The daemon suite has 3 pre-existing
order-dependent flaky tests — daemon_operation_lock_recovers,
remote_runner_broker_redacts_secret_env, legacy_child_recovery — that
fail identically on clean main: 130 passed / 3 failed on both branch and
main; verified by stash. Not introduced here.)
